ref #1130 -- moved around pagination methods
